Fix `logger` issue while running test (#3)
The core change involves leveraging ‘spyOn‘ not just for monitoring
method calls, but also for overriding its implementation. We use ‘spyOn‘
combined with ‘mockImplementation‘ to substitute the original method
with a dummy function for the test's duration, guaranteeing that the
actual method is not invoked. The original implementation is then
restored in the teardown phase.

---------

Signed-off-by: Fery Wardiyanto <ferywardiyanto@gmail.com>